social-story-main-PISC/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
adal==1.2.7
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
async-timeout==4.0.2
attrs==23.1.0
axial-positional-embedding==0.2.1
azure-batch==14.2.0
azure-common==1.1.28
azure-core @ file:///home/conda/feedstock_root/build_artifacts/azure-core_1717713606516/work
azure-identity @ file:///home/conda/feedstock_root/build_artifacts/azure-identity_1719036659591/work
azure-storage-blob==12.21.0
black==23.7.0
blis==0.7.11
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
cachetools==5.3.1
catalogue==2.0.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1696001684923/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
cmake==3.28.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
CoLT5-attention==0.10.20
confection==0.1.4
contourpy==1.2.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1721521267415/work
cycler==0.12.1
cymem==2.0.8
Cython==3.0.0
defusedxml==0.7.1
diffusers==0.26.3
distro==1.9.0
einops==0.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
fairscale==0.4.13
fastapi==0.105.0
filelock==3.13.1
fire==0.5.0
fonttools==4.47.2
frozenlist==1.4.0
fschat==0.2.34
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1666808654411/work
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.2
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna==3.4
imageio==2.34.0
importlib-metadata==7.0.1
iopath==0.1.9
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
lit==17.0.6
local-attention==1.9.0
lvis==0.5.3
Markdown==3.4.3
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==2.1.3
matplotlib==3.7.0
mdurl==0.1.2
mmcv==2.1.0
mmengine==0.10.3
mpmath==1.3.0
msal @ file:///home/conda/feedstock_root/build_artifacts/msal_1721216352388/work
msal-extensions @ file:///home/conda/feedstock_root/build_artifacts/msal_extensions_1652707775195/work
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nh3==0.2.15
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.0
opencv-python==4.8.0.74
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==1.5.3
pathspec==0.11.1
pathy==0.10.2
pillow==10.2.0
plac==1.1.3
platformdirs==4.1.0
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1720928530976/work
preshed==3.0.9
product-key-memory==0.2.10
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1722701264352/work
pyparsing==3.1.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
reformer-pytorch==1.4.4
regex==2023.10.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.12.0
segment-anything==1.0
sentencepiece==0.1.99
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
supervision==0.18.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.1
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.65.0
transformers==4.36.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.25.0
wasabi==1.1.2
wavedrom==2.0.3.post3
wcwidth==0.2.12
weasel==0.3.4
Werkzeug==2.3.6
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
